changed backend endpoint and locales folder
