+++ .github/workflows/azure-webapps-node.yml
@@ -27,8 +27,7 @@ env:
   AZURE_WEBAPP_NAME: unicloud-frontapp # set this to your application's name
   AZURE_WEBAPP_PACKAGE_PATH: 'dist' # set this to the path to your web app project, defaults to the repository root
   NODE_VERSION: '16.x' # set this to the node version to use
-  VITE_BASE_URL: $(VITE_BASE_URL)
-  VITE_PUBLIC_URL: $(VITE_PUBLIC_URL)
+  VITE_BACKEND_URL: ${{ secrets.VITE_BACKEND_URL }}
 
 permissions:
   contents: read
