+++ locales/en/dashboard.json
