+++ locales/es/dashboard.json
