+++ locales/fr/dashboard.json
