+++ locales/fr/login-register.json
