+++ locales/pt/common.json
