+++ locales/pt/customers-users.json
