+++ locales/pt/dashboard.json
