+++ locales/pt/login-register.json
