+++ src/App.tsx
@@ -1,12 +1,7 @@
 import MainRoutes from './routes'
-import { ReactQueryDevtools } from 'react-query/devtools'
+
 const App: React.FC = () => {
-  return (
-    <>
-      <MainRoutes />
-      <ReactQueryDevtools />
-    </>
-  )
+  return <MainRoutes />
 }
 
 export default App
