+++ src/services/api.ts
@@ -8,7 +8,7 @@ let failedRequestQueue: any = []
 let URL_BACKEND = 'https://unicloudbr.azurewebsites.net'
 
 export const api = axios.create({
-  baseURL: URL_BACKEND
+  baseURL: import.meta.env.VITE_BACKEND_URL
 })
 
 api.interceptors.request.use(
